Add placeholder files for chapter 3.
If you compile the book, you'll see there's already a chapter 3. That's because
I decided to devote all of chapter 2 to sets, we'll get to the special types of
sets in chapter 3.

Anyway, I was too lazy to move the chapter 3 text to a new file, so, for the
time being, I just `touch`ed the files, and will use them as placeholders.

I'll move the text tomorrow morning.

Night, everyone.