[en][§3.2] Add subsets/supersets section.
It's a stub, but it's a (re)start

I already wrote a section, but it sucked, so I'm redoing it. This is a good
start to the redo. This version more succinct, and it is much easier to read. It
allows me to get to interesting content much more quickly.

Cheers!