[en] Add commented out section about Printing
As noted in b947268, it should be the case that the print font and the screen
font are different. However, I found a serif font that works for screens - it
looks a bit like Times New Roman. Anyway, I digress. I had originally added a
note telling people to change the font if they were going to print the book. I
removed that note.